Rename java folder to generator
